+++ package.json
@@ -5,7 +5,7 @@
   "main": "app.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "dev": "nodemon npm"
+    "dev": "nodemon app.js"
   },
   "repository": {
     "type": "git",
