+++ public/style.css
@@ -1,0 +1,4 @@
+*{
+    margin: 0;
+    padding: 0;
+}
