+++ views/auth-view/signup.hbs
@@ -1,0 +1,25 @@
+<h1>Signup</h1>
+
+<form action="/signup" method="POST">
+
+  <label for="username">Username*:
+    <input type="text" name="username" id="username" required>
+  </label>
+  
+      <label for="email">e-mail*:
+    <input type="text" name="email" id="email" required>
+  </label>
+  
+  <label for="password">Password*:
+    <input type="text" name="password" id="password" required>
+  </label>
+  
+  {{#if errorMessage}}
+    <div class="error-message">{{ errorMessage }}</div>
+  {{/if}}
+  
+  <button type="submit">Create account</button>
+  
+  <p>Already have an account? <a href="/login">Login</a></p>
+  
+</form>
