+++ views/layout.hbs
@@ -1,0 +1,36 @@
+<!DOCTYPE html>
+<html lang="en">
+
+  <head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <meta http-equiv="X-UA-Compatible" content="ie=edge">
+    <title>{{title}}</title>
+    <link rel="stylesheet" href="/style.css" />
+  </head>
+
+  <body>
+    {{#if currentUser}}
+    <nav>
+      <ul>
+        <li><a href="/">Home</a></li>
+        <li><a href="./user/profile">Perfil</a></li>
+        <li><a href="./plants/list">Plantas</a></li>
+        <li><a href="/logout">Cerrar Sesion</a></li>
+      </ul>
+    </nav>
+    {{else}}
+    <div class="home">
+      <ul>
+        <li><a href="/signup">Registrarse</a></li>
+        <li><a href="/login">Iniciar Session</a></li>
+      </ul>
+      </div>
+    {{/if}}
+
+    {{{body}}}
+
+    <!--<script src="/javascripts/script.js"></script>-->
+  </body>
+
+</html>
