+++ views/plants/list.hbs
@@ -1,0 +1,1 @@
+<h1>lista de plantas</h1>
