+++ views/user/profile.hbs
@@ -1,0 +1,5 @@
+<h1>Private page</h1>
+
+{{#if message}}
+  <p>{{message}} {{user}}</p>
+{{/if}}
